# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PaymentEntity API
  version: 0.2.0
  description: |
    API contract for payment lifecycle operations.
servers:
  - url: /
security:
  - BearerAuth: [ ]
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    PaymentEntity:
      type: object
      required: [ id, status, amount, currency ]
      properties:
        id:
          type: string
          description: ULID payment identifier
        status:
          type: string
          example: processing
        amount:
          type: string
          example: "12.50"
        currency:
          type: string
          example: USD
        providerRef:
          type: string
          nullable: true
          example: stripe_pi_123
    ValidationError:
      type: object
      required: [ field, message ]
      properties:
        field:
          type: string
          example: amount
        message:
          type: string
          example: Use decimal amount format like 50.00.
    ValidationErrorResponse:
      type: object
      required: [ errors ]
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
    ErrorResponse:
      type: object
      required: [ error ]
      properties:
        error:
          type: string
          example: payment-not-found
    StartPaymentRequest:
      type: object
      required: [ amount, currency, provider ]
      properties:
        amount:
          type: string
          example: "12.50"
        currency:
          type: string
          example: USD
        provider:
          type: string
          enum: [ internal, stripe, paypal ]
    CreatePaymentRequest:
      type: object
      required: [ orderId, amountMinor, currency ]
      properties:
        orderId:
          type: string
          example: order-1001
        amountMinor:
          type: integer
          minimum: 1
          example: 1250
        currency:
          type: string
          example: USD
    RefundPaymentRequest:
      type: object
      required: [ amount, provider ]
      properties:
        amount:
          type: string
          example: "10.00"
        provider:
          type: string
          enum: [ internal, stripe, paypal ]
paths:
  /payment/start:
    post:
      summary: Start payment
      security:
        - BearerAuth: [ ]
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartPaymentRequest'
      responses:
        '200':
          description: PaymentEntity started
          content:
            application/json:
              schema:
                type: object
                required: [ payment, provider, status ]
                properties:
                  payment:
                    type: string
                  provider:
                    type: string
                  status:
                    type: string
                  providerRef:
                    type: string
                    nullable: true
                  result:
                    type: object
                    additionalProperties: true
        '400':
          description: Invalid JSON body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Missing or invalid bearer token
        '403':
          description: Missing payment:write scope
        '422':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /api/payments:
    post:
      summary: Create payment
      security:
        - BearerAuth: [ ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePaymentRequest'
      responses:
        '201':
          description: PaymentEntity created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentEntity'
        '400':
          description: Invalid JSON body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Missing or invalid bearer token
        '403':
          description: Missing payment:write scope
        '422':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /api/payments/{id}/refund:
    post:
      summary: Refund payment
      security:
        - BearerAuth: [ ]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefundPaymentRequest'
      responses:
        '200':
          description: PaymentEntity refunded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentEntity'
        '400':
          description: Invalid JSON body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Missing or invalid bearer token
        '403':
          description: Missing payment:write scope
        '404':
          description: PaymentEntity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
  /payment/finalize/{id}:
    post:
      summary: Finalize payment
      security:
        - BearerAuth: [ ]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PaymentEntity finalized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentEntity'
        '401':
          description: Missing or invalid bearer token
        '403':
          description: Missing payment:write scope
        '404':
          description: PaymentEntity not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'